fix(ui): Simplify Sidebar File Structure
Die Komponenten MainSidebar, NavMain und SubNav werden jetzt in der Datei des SidebarLayout definiert, da diese nur von SidebarLayout verwendet werden und ebenfalls als Server-Komponenten gerendert werden. Die Client-Komponenten MainSidebarMenuButton und NavUser werden weiterhin in zwei eigenständigen Dateien gespeichert.
