+++ packages/ui/components/sidebar-layout.tsx
@@ -1,20 +1,37 @@
-import type { ServiceType } from "@northware/service-config";
+import { currentUser } from "@northware/auth/server";
+import { type ServiceType, suiteApps } from "@northware/service-config";
+import { AppSwitch, type MenuApps } from "@northware/ui/components/app-switch";
 import {
   AutoBreadcrumbs,
   type BreadcrumbType,
 } from "@northware/ui/components/auto-breadcrumbs";
 import { buttonVariants } from "@northware/ui/components/button";
-import { MainSidebar } from "@northware/ui/components/main-sidebar";
-import type { SubMenuItem } from "@northware/ui/components/nav-main";
+import {
+  Collapsible,
+  CollapsibleContent,
+} from "@northware/ui/components/collapsible";
 import { Separator } from "@northware/ui/components/separator";
 import {
+  Sidebar,
+  SidebarContent,
+  SidebarFooter,
+  SidebarGroup,
+  SidebarGroupContent,
+  SidebarGroupLabel,
+  SidebarHeader,
   SidebarInset,
+  SidebarMenu,
+  SidebarMenuItem,
+  SidebarMenuSub,
+  SidebarMenuSubItem,
   SidebarProvider,
+  SidebarRail,
   SidebarTrigger,
 } from "@northware/ui/components/sidebar";
+import { MainSidebarMenuButton } from "@northware/ui/components/sidebar-menu-button";
+import { NavUser } from "@northware/ui/components/sidebar-nav-user";
 import { ThemeSwitch } from "@northware/ui/components/theme-switch";
-
-// TODO: Add: Entrypoint parameter, mit dem die Sidebar auch an einem anderen Punkt laut menuData starten kann
+import { menuData } from "@northware/ui/lib/menu-data";
 
 export function SidebarLayout({
   children,
@@ -59,5 +76,158 @@ export function SidebarLayout({
         <div className="flex flex-1 flex-col gap-4 p-4 pt-4">{children}</div>
       </SidebarInset>
     </SidebarProvider>
+  );
+}
+
+interface MainSidebarType extends React.ComponentProps<typeof Sidebar> {
+  service: ServiceType;
+  mainLabel?: string;
+  subLabel?: string;
+  subMenu?: SubMenuItem[];
+}
+
+async function MainSidebar({
+  service,
+  mainLabel,
+  subLabel,
+  subMenu,
+  ...props
+}: MainSidebarType) {
+  const user = await currentUser();
+  const menuItems = await menuData(service, user?.id);
+  const apps: MenuApps[] = suiteApps.map((app) => {
+    const envKey = `NEXT_PUBLIC_${app.slug.toUpperCase()}_URL`;
+    const url = process.env[envKey as keyof typeof process.env];
+
+    return {
+      slug: app.slug,
+      url: url,
+    };
+  });
+
+  return (
+    <Sidebar {...props} variant="inset" collapsible="offcanvas">
+      <SidebarHeader>
+        <AppSwitch service={service} apps={apps} />
+      </SidebarHeader>
+      <SidebarContent>
+        {subMenu && <SubNav subLabel={subLabel} subMenu={subMenu} />}
+        <NavMain menuItems={menuItems} mainLabel={mainLabel} />
+      </SidebarContent>
+      <SidebarFooter>
+        <NavUser
+          fullName={user?.fullName}
+          emailAddress={user?.emailAddresses[0].emailAddress}
+          avatar={user?.imageUrl}
+        />
+      </SidebarFooter>
+      <SidebarRail />
+    </Sidebar>
+  );
+}
+
+type MenuItem = {
+  itemId: string;
+  title: string;
+  href: string;
+  childOf: string | null;
+  permissionKey: string | null;
+};
+
+type MenuType = {
+  topLevelItems: MenuItem[];
+  childItems: (parent: string) => MenuItem[];
+};
+
+function NavMain({
+  menuItems,
+  mainLabel,
+}: { menuItems: MenuType; mainLabel?: string }) {
+  return (
+    <SidebarGroup>
+      {mainLabel && <SidebarGroupLabel>{mainLabel}</SidebarGroupLabel>}
+      <SidebarGroupContent>
+        <SidebarMenu>
+          <SidebarMenuItem>
+            <MainSidebarMenuButton href="/" title="Home" type="topLevel" />
+          </SidebarMenuItem>
+          {menuItems.topLevelItems.map((item) => {
+            const ItemChildren = menuItems.childItems(item.itemId);
+            if (ItemChildren.length === 0) {
+              // Top-Level-Element ohne Unterpunkte
+              return (
+                <SidebarMenuItem key={item.itemId}>
+                  <MainSidebarMenuButton
+                    href={item.href}
+                    title={item.title}
+                    type="topLevel"
+                  />
+                </SidebarMenuItem>
+              );
+            }
+
+            return (
+              // Top-Level-Element mit Unterpunkten
+              <Collapsible
+                key={item.itemId}
+                asChild
+                className="group/collapsible"
+              >
+                <SidebarMenuItem>
+                  <MainSidebarMenuButton
+                    title={item.title}
+                    type="parent"
+                    href={item.href}
+                  />
+                  <CollapsibleContent>
+                    <SidebarMenuSub>
+                      {ItemChildren.map((subItem) => (
+                        <SidebarMenuSubItem key={subItem.itemId}>
+                          <MainSidebarMenuButton
+                            title={subItem.title}
+                            href={subItem.href}
+                            type="child"
+                          />
+                        </SidebarMenuSubItem>
+                      ))}
+                    </SidebarMenuSub>
+                  </CollapsibleContent>
+                </SidebarMenuItem>
+              </Collapsible>
+            );
+          })}
+        </SidebarMenu>
+      </SidebarGroupContent>
+    </SidebarGroup>
+  );
+}
+
+type SubMenuItem = {
+  title: string;
+  href: string;
+};
+
+function SubNav({
+  subLabel,
+  subMenu,
+}: { subLabel?: string; subMenu: SubMenuItem[] }) {
+  return (
+    <SidebarGroup>
+      <SidebarGroupLabel>{subLabel || "Seitennavigation"}</SidebarGroupLabel>
+      <SidebarGroupContent>
+        <SidebarMenu>
+          <SidebarMenuItem>
+            {subMenu.map((item) => (
+              <MainSidebarMenuButton
+                key={item.title}
+                href={item.href}
+                title={item.title}
+                type="topLevel"
+              />
+            ))}
+          </SidebarMenuItem>
+        </SidebarMenu>
+      </SidebarGroupContent>
+    </SidebarGroup>
   );
 }
